authorized route: staff,service, room. Create route: login, myprofile

+++ src/Hotel.API/Controllers/StaffController.cs
@@ -8,6 +8,7 @@ using System.Security.Claims;
 
 namespace Hotel.API.Controllers;
 
+[Authorize(Roles = "Manager")]
 [Route("api/[controller]")]
 [ApiController]
 public class StaffController : ControllerBase
@@ -26,52 +27,51 @@ public class StaffController : ControllerBase
 
     public async Task<ActionResult> Get()
     {
-        return Ok(await _userService.GetUserListAsync());
+        var result = await _userService.GetUserListAsync();
+        return Ok(result);
     }
 
     [HttpGet("search")]
     public async Task<ActionResult<UserToReturnDTO>> Search([FromQuery] string value, string option = "id")
     {
 
-        if (option.IsNullOrEmpty() || value.IsNullOrEmpty()) return BadRequest();
-
-        return Ok(await _userService.SearchUserAsync(option, value));
+        if (option.IsNullOrEmpty() || value.IsNullOrEmpty()) return BadRequest("Invalid request.");
+        var result = await _userService.SearchUserAsync(option, value);
+        return Ok(result);
     }
 
     [HttpPost("")]
     public async Task<ActionResult> CreateUser([FromBody] UserToCreateDTO userToCreateDTO)
     {
-        return Ok(await _userService.CreateUserAsync(userToCreateDTO));
-    }
-
-    [Authorize(Roles = "staff")]
-    [HttpGet("profile")]
-    public async Task<ActionResult> GetUserDetailByID([FromQuery] int id)
-    {
-        // var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
-        var result = await _userService.GetUserByIDAsync(id);
+        var result = await _userService.CreateUserAsync(userToCreateDTO);
         return Ok(result);
     }
 
-    [HttpPost("login")]
-    public async Task<ActionResult> Login(UserCreateDto user)
+   
+    [HttpGet("{username}")]
+    public async Task<ActionResult> GetUserDetailByUsername(string username)
     {
-        (var token, var userDto) = await _userService.Login(user);
-        return Ok(new { Token = token, user = userDto });
+        var result = await _userService.GetUserByUsernameAsync(username);
+        return Ok(result);
     }
 
     [HttpPut("")]
     public async Task<ActionResult> ChangeUserPassword([FromQuery] int id, [FromBody] string newPassWord)
     {
         await _userService.ChangeUserPassWordAsync(id, newPassWord);
-        return Ok("ok");
+        return Ok($"Changed password: User #'{id}'.");
     }
 
     [HttpDelete("")]
     public async Task<ActionResult> RemoveUser([FromQuery] int id)
     {
+        var currentUserID = Int32.Parse(User.FindFirstValue(ClaimTypes.NameIdentifier)!);
+        if (id== currentUserID)
+        {
+            return BadRequest("Cannot remove self.");
+        }
         await _userService.RemoveUserAsync(id);
-        return Ok($"Removed user #'{id}'.");
+        return Ok($"Removed: User #'{id}'.");
     }
 }
